refactor(settings page & language selector): separate language selector
